+++ src/components/CarrinhoVazio.css
@@ -1,0 +1,5 @@
+.vazio {
+  font-size: 80px;
+  margin-top: 150px;
+  text-align: center;
+}
