+++ src/components/Header.css
@@ -10,12 +10,20 @@
 }
 
 .inputBusca {
-  border-radius: 5%;
-  width: 100%;
+  border-color: rgb(235, 235, 235);
+  border-radius: 5px;
+  height: 25px;
+  width: 600px;
 }
 
 .botao {
   align-items: flex-end;
   height: 20px;
   margin: 35px 5px 0;
+}
+
+.menu-nav {
+  background-color: #fff159;
+  border-color: #fff159;
+  width: 180px;
 }
