+++ src/page/Cart.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import CartVazio from '../components/CarrinhoVazio';
+
+export default class Cart extends Component {
+  render() {
+    return (
+      <div>
+        <CartVazio />
+      </div>
+    );
+  }
+}
